Create PrimeNumbersBetweenTwoIntervals.cpp
To print all prime numbers between two integers, checkPrimeNumber() function is created. This function checks whether a number is prime or not.

All integers between n1 and n2 are passed to this function.

If a number passed to checkPrimeNumber() is a prime number, this function returns true, if not the function returns false.

If the user enters the larger number first, this program will swap the numbers. Without swapping, this program won't work.