Create ConversionExample1IntToDouble.cpp
Output

num_int = 9
num_double = 9
In the program, we have assigned an int data to a double variable.

num_double = num_int;
Here, the int value is automatically converted to double by the compiler before it is assigned to the num_double variable. This is an example of implicit type conversion.