Create ConversionExample2DoubleToInt.cpp
Output

num_int = 9
num_double = 9.99
In the program, we have assigned a double data to an int variable.

num_double = num_int;
Here, the double value is automatically converted to int by the compiler before it is assigned to the num_int variable. This is also an example of implicit type conversion.

Note: Since int cannot have a decimal part, the digits after the decimal point are truncated in the above example.