Create BasicIoNumbersAndCharactersOutput.cpp
Output

70
256.783
character: A
Notes:

The endl manipulator is used to insert a new line. That's why each output is displayed in a new line.
The << operator can be used more than once if we want to print different variables, strings and so on in a single statement. For example:
cout << "character: " << ch << endl;