Create BasicIoInputOutput
Output

Enter an integer: 70
The number is: 70
In the program, we used

cin >> num;
to take input from the user. The input is stored in the variable num. We use the >> operator with cin to take input.

Note: If we don't include the using namespace std; statement, we need to use std::cin instead of cin.