Create ArraySumAndAvarageArrayElementsUsingForLoop.cpp
Output

The numbers are: 7  5  6  12  35  27
Their Sum = 92
Their Average = 15.3333
In this program:

We have initialized a double array named numbers but without specifying its size. We also declared three double variables sum, count, and average.

Here, sum =0 and count = 0.
Then we used a range based for loop to print the array elements. In each iteration of the loop, we add the current array element to sum.
We also increase the value of count by 1 in each iteration, so that we can get the size of the array by the end of the for loop.
After printing all the elements, we print the sum and the average of all the numbers. The average of the numbers is given by average = sum / count;
Note: We used a ranged for loop instead of a normal for loop.

A normal for loop requires us to specify the number of iterations, which is given by the size of the array.

But a ranged for loop does not require such specifications.