Create CheckOddEven_IfElse.cpp
Output

Enter an integer: 23
23 is odd.
In this program, if..else statement is used to check whether n%2 == 0 is true or not. If this expression is true, n is even if not n is odd.

You can also use ternary operators ?: instead of if..else statement. Ternary operator is short hand notation of if...else statement.